django_app/requirements.txt
Django>=4.0,<5.0
djangorestframework
mysqlclient
gunicorn
pyyaml
requests
django-cors-headers
python-dotenv
pycurl